Rename SchedulerLoop -> Controller
